refactor query_profile table
