+++ src/query/service/src/interpreters/interpreter_delete.rs
@@ -366,7 +366,6 @@ pub async fn subquery_filter(
         &[row_id_column_binding.clone()],
         &physical_plan,
         false,
-        false,
     )
     .await?;
