+++ src/query/storages/system/src/lib.rs
@@ -43,6 +43,7 @@ mod processes_table;
 mod query_cache_table;
 mod query_log_table;
 mod query_profile_table;
+mod query_summary_table;
 mod roles_table;
 mod settings_table;
 mod stages_table;
@@ -86,6 +87,7 @@ pub use query_log_table::QueryLogElement;
 pub use query_log_table::QueryLogQueue;
 pub use query_log_table::QueryLogTable;
 pub use query_profile_table::QueryProfileTable;
+pub use query_summary_table::QuerySummaryTable;
 pub use roles_table::RolesTable;
 pub use settings_table::SettingsTable;
 pub use stages_table::StagesTable;
